% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_artifact.R
\name{download_ci_artifact}
\alias{download_ci_artifact}
\title{Download the latest snapshot artifacts from the CI and update local snapshots}
\usage{
download_ci_artifact(
  name,
  to = ".",
  branch = get_branch(),
  repo = get_repo(),
  owner = get_owner()
)
}
\arguments{
\item{name}{The name of the artifact}

\item{to}{Output directory for the artifact}

\item{branch}{The name of the branch}

\item{repo}{The name of the repository}

\item{owner}{The name of the owner of the repository}
}
\value{
NULL, called for side effects
}
\description{
Download the latest snapshot artifacts from the CI and update local snapshots
}
\details{
This function uses Github API to download the latest snapshot artifacts from the CI and update the local snapshots.
}
