% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{git}
\alias{git}
\alias{get_branch}
\alias{get_repo}
\alias{get_owner}
\title{Helper functions to quickly target the remote repository}
\usage{
get_branch()

get_repo()

get_owner()
}
\description{
Helper functions to quickly target the remote repository
}
